requirements.txt
pyyaml==6.0.1
pandas==2.1.4
scikit-surprise==1.1.3
mlflow==2.12.2